Update README.md
Please, just do a favour and install all of the app files and not just the index.html file. All of the files are for Visual Studio Code. Just reminding you again: Make sure you have Windows, (please do not try this on macOS) Visual Studio Code and post importantly-Google Chrome. Still, all of these are important and it won't be able to use it without Windows all of these apps I mentioned.